EDX-2853 - switches sldmove to use sdc api EDX-2869 - shows merged data on sld history and compare screens
