+++ frontend/src/components/CompareStudents.vue
@@ -5,35 +5,35 @@
         id="compareClearBtn"
         text="Clear"
         secondary
-        @click-action="[cancel, clearError]"
+        @click-action="[cancel(), clearError()]"
       />
       <PrimaryButton
         id="moveSldBtn"
         text="Move Sld Record"
         primary
-        :disabled="disableMoveSld"
-        @click-action="moveSldRecords"
+        :disabled="disableMoveSld()"
+        @click-action="moveSldRecords()"
       />
       <PrimaryButton
         id="twinBtn"
         text="Twin"
         primary
-        :disabled="validateAction"
-        @click-action="twin"
+        :disabled="validateAction()"
+        @click-action="twin()"
       />
       <PrimaryButton
         id="demergeBtn"
         text="Demerge"
         primary
         :disabled="disableDemerge()"
-        @click-action="demerge"
+        @click-action="demerge()"
       />
       <PrimaryButton
         id="mergeBtn"
         text="Merge PENs"
         primary
         :disabled="disableMerge()"
-        @click-action="merge"
+        @click-action="merge()"
       />
     </template>
   </CompareDemographicsCommon>
